+++ frontend/src/pages/auth/sign-up.tsx
@@ -1,7 +1,7 @@
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { Link, useNavigate } from "react-router-dom";
-import { useAuth } from "@/context/AuthContext";
+import { useAuth } from "@/src/context/AuthContext";
 import { signUpSchema, type SignUpFormData } from "@/src/lib/validations/auth";
 import { FormInput } from "@/src/components/ui/!to-migrate/form-input";
 import { Button } from "@/src/components/ui/!to-migrate/button";
